Project (AI & GenAI)/Project/Project/modules/embed_store.py
import faiss
import numpy as np
from sentence_transformers import SentenceTransformer

def embed_and_store(chunks):
    model = SentenceTransformer('all-MiniLM-L6-v2')
    embeddings = model.encode(chunks)
    dim = embeddings.shape[1]
    index = faiss.IndexFlatL2(dim)
    index.add(np.array(embeddings))
    return index, model, chunks
