Project (AI & GenAI)/Project/Project/requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
streamlit


